requirements.txt
# Chat web (Chainlit)
chainlit==2.11.1

# LLM — Groq via OpenAI SDK
openai==2.32.0

# Variables de entorno
python-dotenv==1.2.2

# Google Sheets
gspread==6.2.1
google-auth==2.49.2
google-auth-oauthlib==1.3.1

# API de voz (Vapi backend)
fastapi==0.136.0
uvicorn==0.45.0
pydantic==2.13.3

# HTTP client (tests y FastAPI TestClient)
httpx==0.28.1

# Detección de idioma
langdetect==1.0.9
